add connection between two pages
