+++ index.html
@@ -71,12 +71,12 @@
                       <div class="form-group">
                         <label for="input-from-flights">From</label>
                         <input type="text" class="form-control form-control-sm" id="input-from-flights"
-                          placeholder="Minsk">
+                          placeholder="Minsk" required>
                       </div>
                       <div class="form-group">
                         <label for="input-to-flights">To</label>
                         <input type="text" class="form-control form-control-sm" id="input-to-flights"
-                          placeholder="Moscow">
+                          placeholder="Moscow" required>
                       </div>
                       <div class="d-flex justify-content-end">
                         <button type="submit" class="btn btn-primary search" id="search-flights"
@@ -201,7 +201,7 @@
   <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.2.1/js/bootstrap.min.js"
     integrity="sha384-B0UglyR+jN6CkvvICOB2joaf5I4l3gm9GU6Hc1og6Ls7i6U/mkkaduKaBhlAXv9k"
     crossorigin="anonymous"></script>
-  <script src="dist/bundle.js"></script>
+  <script src="dist/home.js"></script>
 </body>
 
 </html>
